requirements.txt
pandas==2.3.0
Pillow==11.2.1
pytesseract==0.3.10
